requirements.txt
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2
groq==0.4.1
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
streamlit==1.29.0
selenium==4.15.2
lxml==4.9.3